requirements.txt
# Core Dependencies
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.65.0

# Data Analysis and Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Financial Data and Analysis
yfinance>=0.2.36
openpyxl>=3.1.2
xlsxwriter>=3.0.0

# Export and Documentation
reportlab>=3.6.0
weasyprint>=60.1

# News and Social Media Analysis
feedparser>=6.0.10
spacy>=3.0.0
transformers>=4.0.0
tweepy>=4.14.0
praw>=7.7.0

# AI and Machine Learning
openai>=1.0.0

# Testing
pytest>=7.0.0
